cilium: Use correct mask on xfrm delete
The mask values used to match stale entries in the XFRM state and policy
no longer matched the values we use. The result is stale entries get
left in the node tables. If subsequently cilium agent picks a new IP
address this can result in conflicts in the xfrm {state|policy} lookup

Fix to use correct mask field.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>