test: Make Wireguard tcpdump filter more fine grained
It was reported, that the native Wireguard CI test failed with the
tcpdump assertion. This means that the tcpdump was able to capture
a packet on the private iface (used for direct routing) targeting
the pod on any port.

Improve the filter by adding 'port $targetPort' filter to eliminate non
service traffic. Also, write tcpdump output into test-output.log, so
that next time we know whether a failure was false postivie.

Reported-by: Nate Sweet <nathanjsweet@pm.me>
Signed-off-by: Martynas Pumputis <m@lambda.lt>