pkg/endpoint: derive endpoint from SpiffeID
This code creates a new goroutime when "SetPod" is called. This
goroutime uses the helpers introduced in the previous commit to get the
SpiffeID of this endpoint, once the ID is known, the security labels
of the endpoint are updated by introducing a new
"spiffe:spiffeID=...." label.

To conditionals are introduced in ModifyIdentityLabels & UpdateLabels
to avoid overwriting the security labels of an endpoint that has its
spiffeID calculated.