envoy: create HTTP to HTTPS filter configuration
Add a filter chain that matches HTTP traffic and uses an HTTPS egress
cluster.

TODO: This commit probably breaks other network policies as all HTTP
traffic will be redirected to cluster that uses TLS.
Maybe we can do the "cilium.tls_wrapper" implementation in cilium-proxy
smart enough to avoid using TLS when it's not actually required or
should we implement that logic on cilium-agent to select the right
TransportSocket according to the policy configuration?